Fix project name; add constant_time module; add .gitignore
The pqcrypto_sidh crate name is supposed to be snake case.

Add constant_time::is_nonzero and constant_time::is_zero with implementations and tests, and constant_time::is_less_than with a stub and a failing test.

Add a .gitignore file to blacklist the target build and lock files from commits.